Create notification_badges.sh
Bash script which does install ubuntu-keyring in order to install libindicator7 from Ubuntu 20.04 LTS repository. It also install dash-to-dock extension and enables it. The outcome is getting notification badges for example on the dock icons like Thunderbird, Nautilus, etc. It's reasonably safe because there is no complex dendencies nad the Ubuntu repositories are removed after installation to avoid "Hell Of Dependencies".